separated spec files for pricing box
