+++ cypress/integration/v2/blocks/pricing-box/tablet-advanced.spec.js
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import { tabletAdvanced } from '.'
+
+describe( 'Pricing Box Block ( Tablet Advanced )', registerTests( [ tabletAdvanced ] ) )
